+++ src/providers/generalWalletProvider.tsx
@@ -17,12 +17,6 @@ import {
 	SystemProgram,
 } from '@solana/web3.js';
 import {
-	createAssociatedTokenAccountInstruction,
-	createTransferInstruction,
-	getAssociatedTokenAddress,
-	TOKEN_PROGRAM_ID,
-} from '@solana/spl-token';
-import {
 	Chain,
 	useAccount,
 	useBalance,
@@ -65,11 +59,6 @@ interface IGeneralWalletContext {
 		to: string,
 		value: string,
 	) => Promise<string | `0x${string}` | undefined>;
-	sendSolanaSPLToken: (
-		to: string,
-		value: bigint,
-		tokenAddress: string,
-	) => Promise<string | undefined>;
 	handleSingOutAndSignInWithEVM: () => Promise<void>;
 	handleSignOutAndSignInWithSolana: () => Promise<void>;
 	handleSignOutAndShowWelcomModal: () => Promise<void>;
@@ -88,7 +77,6 @@ export const GeneralWalletContext = createContext<IGeneralWalletContext>({
 	chain: undefined,
 	openWalletConnectModal: () => {},
 	sendNativeToken: async () => undefined,
-	sendSolanaSPLToken: async () => Promise.resolve(''),
 	handleSingOutAndSignInWithEVM: async () => {},
 	handleSignOutAndSignInWithSolana: async () => {},
 	handleSignOutAndShowWelcomModal: async () => {},
@@ -324,50 +312,6 @@ export const GeneralWalletProvider: React.FC<{
 		}
 	};
 
-	const sendSolanaSPLToken = async (
-		to: string,
-		value: bigint,
-		tokenAddress: string,
-	) => {
-		if (!publicKey) throw Error('Wallet is not connected');
-		const splTokenMintAddress = new PublicKey(tokenAddress);
-		const receiverAddress = new PublicKey(to);
-		const senderTokenAccountAddress = await getAssociatedTokenAddress(
-			splTokenMintAddress,
-			publicKey,
-		);
-		const receiverTokenAccountAddress = await getAssociatedTokenAddress(
-			splTokenMintAddress,
-			receiverAddress,
-		);
-		const transaction = new Transaction();
-		const receiverAccountInfo = await solanaConnection.getAccountInfo(
-			receiverTokenAccountAddress,
-		);
-		if (!receiverAccountInfo) {
-			// In the case where user is new to the token and doesn't have an associated token account
-			transaction.add(
-				createAssociatedTokenAccountInstruction(
-					publicKey,
-					receiverTokenAccountAddress,
-					receiverAddress,
-					splTokenMintAddress,
-				),
-			);
-		}
-		transaction.add(
-			createTransferInstruction(
-				senderTokenAccountAddress,
-				receiverTokenAccountAddress,
-				publicKey,
-				value,
-				[],
-				TOKEN_PROGRAM_ID,
-			),
-		);
-		return await solanaSendTransaction(transaction, solanaConnection);
-	};
-
 	const disconnect = () => {
 		switch (walletChainType) {
 			case ChainType.EVM:
@@ -462,7 +406,6 @@ export const GeneralWalletProvider: React.FC<{
 		openWalletConnectModal,
 		balance,
 		sendNativeToken,
-		sendSolanaSPLToken,
 		handleSingOutAndSignInWithEVM,
 		handleSignOutAndSignInWithSolana,
 		handleSignOutAndShowWelcomModal,
